Add basic logging functionality.
The README file is updated to indicate an installation requirement for "log4r".

The Analysis.R file updated to write log messages for each step taken in that file.

Furthermore Analysis.R is updated so that data initialization is performed exclusively with targeted functions, rather than with one over-arching function. Log messages indicate the completion of each initialization step.

Finally, Analysis.R is rearranged so that correlation analysis is performed last.  The reason for this is that performance difficulties were observed on Windows systems for this step; but all other steps are reasonably performant on both Linux and Windows.

The major feature of the logging functionality is that each step in Analysis.R is noted, both in console output and in a file, alongside a datetimestamp.  This is useful for observing the execution time required for each step of the script.