Provide changes necessary for correct execution under the Windows Subsystem for Linux (WSL).
During investigation of performance problems on Windows, an attempt was made to execute the package with a Linux-native R binary under WSL.

These attempts led to complaints by KEGG-related code that it could not download and make use of data from a remote website.  The complaints were defeated by specifying the use of "wget" rather than "curl" for the download, and by requiring Reactome packages to be present for the execution of the analysis script.

These changes have been observed to work with WSL, but should be separately evaluated on Linux.